requirements.txt
pytorch==1.13.1
transformers==4.39.3
openai==0.28.0
jinja2==3.1.3